Move @Beans into separate config file
