+++ src/main/java/library_project/Application.java
@@ -12,6 +12,8 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.annotation.AnnotationConfigApplicationContext;
 import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Import;
+import org.springframework.test.context.ContextConfiguration;
 
 import java.sql.Connection;
 import java.sql.DriverManager;
@@ -19,24 +21,8 @@ import java.sql.SQLException;
 import java.util.Properties;
 
 @SpringBootApplication
+@Import(LibraryConfig.class)
 public class Application {
-
-    @Bean
-    public Repository<Reservation> reservationRepoDatabase() {
-        return new ReservationRepoDatabase(databaseConnector());
-    }
-
-    @Bean
-    public Repository<Book> bookRepoDatabase() {
-        return new BookRepoDatabase(databaseConnector());
-    }
-
-    @Bean
-    public DatabaseConnector databaseConnector() {
-        return new DatabaseConnector("jdbc:mysql://localhost:3306/test_schema", "root", "1234");
-    }
-
-
 
     public static void main(String[] args) {
         SpringApplication.run(Application.class, args);
